% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_candidates.R
\name{all_candidates}
\alias{all_candidates}
\title{Candidates Master metadata function}
\usage{
all_candidates(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?candidates}.
}
\description{
\code{all_candidates} returns a dataframe about the candidates master data
}
\examples{
all_candidates()
all_cadidates(n_max = 250)
}
