% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_committees.R
\name{all_committees}
\alias{all_committees}
\title{Committees metadata}
\usage{
all_committees(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?committees}.
}
\description{
\code{all_committees} returns a dataframe about the committees master data
}
\examples{
\dontrun{all_committees()}
\dontrun{all_committees(n_max = 250)}
}
