% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_contributions.R
\name{all_contributions}
\alias{all_contributions}
\title{Committee contributions metadata}
\usage{
all_contributions(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?contributions}.
}
\description{
\code{all_contributions} returns a dataframe about the committees master data
}
\examples{
\dontrun{all_contributions()}
\dontrun{all_contributions(n_max = 250)}
}
