% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_individuals.R
\name{all_individuals}
\alias{all_individuals}
\title{Individual Contributions Master metadata}
\usage{
all_individuals(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?individuals}.
}
\description{
\code{all_individuals} returns a dataframe about the committees master data
}
\examples{
all_individuals()
all_individuals(n_max = 250)
}
