% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_pac.R
\name{all_pac}
\alias{all_pac}
\title{Political Action Committee (PAC) and Party Summary Financial Information}
\usage{
all_pac(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?pac}.
}
\description{
\code{all_pac} returns a dataframe about the committees master data
}
\examples{
\dontrun{all_pac()}
\dontrun{all_pac(n_max = 250)}
}
