% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaigns}
\alias{campaigns}
\title{House/Senate Current Campaigns}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1898 rows and 25 columns.}
\usage{
campaigns
}
\description{
@description Summary financial information for each campaign.
}
\details{

}
\keyword{datasets}
