% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec16-package.R
\docType{package}
\name{fec16}
\alias{fec16}
\title{fec16: A Data Package for the 2016 Election}
\description{
The fec16 package provides Federal Election Commission data of:
candidates, committees, results, individual contributions, and committee contributions.
}
\section{Datasets}{

\if{html}{\figure{hex_fec16.png}{options: align = "right" width="45\%" alt="Figure: hex_fec16.png"}}
\itemize{
\item Candidates: \link{candidates}
\item Committees: \link{committees}
\item House Results: \link{house_results}
\item Senate Results: \link{senate_results}
\item Presidential Results: \link{president_results}
\item Individual Contributions: \link{individuals}
\item Committee Contributions: \link{contributions}
\item House/Senate Current Campaigns: \link{campaigns}
\item Operating Expenditures: \link{expenditures}
\item Transactions From One Committee To Another: \link{transactions}
\item PAC and Party Summary Financial Information: \link{pac}
}
}

\section{Authors}{

\itemize{
\item Irene Ryan: \url{https://github.com/ireneryan}
\item Marium Tapal: \url{https://github.com/mariumtapal}
\item Rana Gahwagy: \url{https://github.com/ranawg}
\item Ben Baumer: \url{https://github.com/beanumber}
}
}

