% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec20-package.R
\docType{package}
\name{fec20}
\alias{fec20}
\title{fec20: A Data Package for the 2020 Elections}
\description{
The fec20 package provides the following 6 datasets from the Federal Election Commission's 2019-2020 election cycle:
}
\section{Datasets}{

\if{html}{\figure{hex_fec20.png}{options: align = "right" width="45\%" alt="Figure: hex_fec16.png"}}
\itemize{
\item Candidates: \link{candidates}
\item Committees: \link{committees}
\item Individual Contributions: \link{individuals}
\item Committee Contributions: \link{contributions}
\item House/Senate Current Campaigns: \link{campaigns}
\item Operating Expenditures: \link{expenditures}
\item Transactions From One Committee To Another: \link{transactions}
\item Political Action Committee (PAC) and Party Summary Financial Information: \link{pac}}

The entire \link{candidates}, \link{committees}, \link{campaigns}, and \link{pac} datasets and samples containing 1000 entries of the other datasets are available directly in the package.
For the datasets containing samples, full versions can be retrieved via the following functions:
\itemize{
\item Individual Contributions: \link{read_all_individuals}
\item Committee Contributions: \link{read_all_contributions}
\item Operating Expenditures: \link{read_all_expenditures}
\item Transactions From One Committee To Another: \link{read_all_transactions}
}

In addition, \code{fec20} provides the \link{states} dataset giving geographical information about the 50 states.

Note: There are 3 forthcoming datasets that contain results from the House, Senate, and Presidential elections.
}

\section{Authors}{

\itemize{
\item Marium Tapal: \url{https://github.com/mariumtapal}
\item Irene Ryan: \url{https://github.com/ireneryan}
\item Rana Gahwagy: \url{https://github.com/ranawg}
\item Ben Baumer: \url{https://github.com/beanumber}
}
}

