% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_results}
\alias{house_results}
\alias{senate_results}
\alias{president_results}
\title{General Federal Election Results}
\format{A data frame with 2110 rows and 10 columns, providing information
for results in each state:
\describe{
\item{cand_id}{A code assigned to a candidate by the FEC}
\item{state}{The state of the votes}
\item{general_election_date}{The election date}
\item{name_first}{First name of candidate}
\item{name_last}{Last name of candidate}
\item{last_name_first}{First and last name of candidate}
\item{party}{Party of the candidates}
\item{general_results}{The number of people voted in that state}
\item{general_percent}{The percent of people}
\item{won}{A logical variable: TRUE if candidate won, FALSE if not}
}}
\source{
\url{https://transition.fec.gov/general/FederalElections2016.shtml}, last edited Feb 2018
}
\usage{
house_results

senate_results

president_results
}
\description{
2016 House General Election Results.

2016 House General Election Results.

2016 Presidential General Election Results.
}
\keyword{datasets}
