% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_house}
\alias{results_house}
\title{House Election Results}
\format{
A data frame with 2526 rows and 13 columns providing information the House election results:
\describe{
\item{state}{The state of the votes}
\item{district_id}{District ID}
\item{cand_id}{A code assigned to a candidate by the FEC}
\item{incumbent}{A logical variable: TRUE if incumbent, FALSE if not}
\item{party}{Party of the candidate}
\item{primary_votes}{Number of votes in the primary election}
\item{primary_percent}{Percentage of votes in the primary election}
\item{runoff_votes}{Number of runoff votes}
\item{runoff_percent}{Percentage of runoff votes}
\item{general_votes}{Number of votes in the general election}
\item{general_percent}{Percentage of votes in the general election}
\item{won}{A logical variable: TRUE if candidate won, FALSE if not}
\item{footnotes}{Footnotes}
}
}
\source{
\url{https://www.fec.gov/introduction-campaign-finance/election-results-and-voting-information/}, updated 2020-03-02
}
\usage{
results_house
}
\description{
2020 House General Election Results
}
\seealso{
\link{results_senate}, \link{results_president}
}
\keyword{datasets}
