% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{US States lookup table}
\format{
A data frame with 50 rows and 3 columns:
\describe{
\item{state}{two-character postal abbreviation}
\item{state_name}{the name of the state}
\item{state_region}{coarse region of the country that the state belongs to}
\item{state_region}{finer region of the country that the state belongs to}
}
}
\source{
\code{\link[datasets]{state}}
}
\usage{
states
}
\description{
US States lookup table
}
\keyword{datasets}
