% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_house}
\alias{results_house}
\alias{results_senate}
\alias{results_president}
\title{General Federal Election Results}
\format{A data frame with 2110 rows and 15 columns, providing information
for results in each state:
\describe{
\item{state}{The state of the votes}
\item{district_id}{District ID}
\item{cand_id}{A code assigned to a candidate by the FEC}
\item{incumbent}{A logical variable: TRUE if incumbent, FALSE if not}
\item{candidate_name_first}{Candidate's First Name}
\item{candidate_name_last}{Candidates's Last Name}
\item{party}{Party of the candidate}
\item{primary_votes}{Number of votes in the primary electiom}
\item{primary_percent}{Percentage of votes in the primary election}
\item{runoff_votes}{Number of runoff votes}
\item{runoff_percent}{Percentage of runoff votes}
\item{general_votes}{Number of votes in the general election}
\item{general_percent}{Percentage of votes in the general election}
\item{general_election_date}{The election date}
\item{name_first}{First name of candidate}
\item{name_last}{Last name of candidate}
\item{last_name_first}{First and last name of candidate}
\item{party}{Party of the candidates}
\item{general_results}{The number of people voted in that state}
\item{general_percent}{The percent of people}
\item{won}{A logical variable: TRUE if candidate won, FALSE if not}
\item{footnotes}{Footnotes}
}}
\source{
\url{https://transition.fec.gov/general/FederalElections2016.shtml}, updated 2020-03-02
}
\usage{
results_house

results_senate

results_president
}
\description{
2016 House General Election Results

2016 Senate General Election Results

2016 Presidential General Election Results
}
\keyword{datasets}
