% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_president}
\alias{results_president}
\title{Presidential Election Results}
\format{
A data frame with 619 rows and 6 columns providing information the House election results:
\describe{
\item{cand_id}{A code assigned to a candidate by the FEC}
\item{state}{The state of the votes}
\item{party}{Party of the candidate}
\item{general_votes}{Number of votes in the general election}
\item{general_percent}{Percentage of votes in the general election}
\item{won}{A logical variable: TRUE if candidate won, FALSE if not}
}
}
\source{
\url{https://transition.fec.gov/general/FederalElections2016.shtml}, updated 2020-03-02
}
\usage{
results_president
}
\description{
2016 House General Election Results
}
\seealso{
\link{results_house}, \link{results_senate}
}
\keyword{datasets}
